feat: create dashboard resources using Terraform
Instead of using the null_resource provider, use the official Kubernetes
resources for provisioning the metrics server and dashboard deployments
in the cluster.

Using Terraform to manage the life cycle of these resources makes them
stateful and easier to tear down the cluster due to the implicit
dependencies that Terraform creates when running the apply operation.